pyproject.toml
[tool.poetry]
name = "stoier"
version = "0.1.0"
description = ""
authors = ["Florian Rämisch <olf@subsignal.org>"]
license = "GPL-3.0"

[tool.poetry.scripts]
00_csv_to_yml = 'stoier.csvtoyaml:csv_to_yml'
01_clean = 'stoier.clean:clean'
02_deduplicate = 'stoier.deduplicate:deduplicate'
03_validate = 'stoier.validate:validate'
04_iterate = 'stoier.iterate:iterate'
05_account = 'stoier.account:account'
06_report = 'stoier.report:report'

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
click = "^8.0.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
